t1.py: switch off soflimits may block the motor
Following situation:
DHLM (low soft limit) = -140
DLLM (high soft limit) = 13
DVAL (motor dial position) = -134

When the soft limits should be disabled, both DLLM and DHLM
must be set to 0.0.
In the ideal world, this should be an atomic operation,
but EPICS PV's don't work this way,
When we set DLLM to 0.0, and then DHLM,
the motor may temporally go into a soft limit violation,
especially if we send the soft limits to the motion controll unit.

Avoid this glitch by setting DLLM to 0.0 first, whenever DVAL > 0,
or  DHLM to 0.0 first, whenever DVAL <= 0,