Support for MRES in the driver
Typically the controller uses engineering units, and the driver
gets "steps" from the motor record.
Introduce a new record to store the motor resolution.
Currently it is set to 0.0046875
The motor record (?) rounds it to 0.004687

Remove m2..m3 from the template

Add Records for offset and direction as well, they are not used at the moment