Add install-epics.sh
To be able to run the epicsIOC on a non-ESS build system, provide
an install-epics.sh script.
The script will download and compile EPICS base, ASYN and motor.

Change install-epics.sh, if you want different version to be installed.
The script has been tested under Linux and Mac OS X (and both can share
the installation)