Error handling: Don't use PROBLEM bit
When the axis can't move as commandet, the bError bit is set.
Currently the bError is forwared into the motoStatusProblem bit.
This has an unwanted side effect:
Whenever the "Problem" bit is the, the motor record sends a stop command,
until the "Problem" bit is cleared.
(This may be needed for MCU's that don't stop on an error)
In our cas we want the error to be on the screen, until the user has pressed
STOP in the motor record.