Error handling: Improve simulator
The new error reporting from MCU makes it possible, to test the error
handling in a more polite way:
Disable the amplifier, and try to JOG.
This leads to an error reported by the MCU.

Adopt the simulator to it.
Modify the axis driver, not to call resetAxis() when amplifier is
switched off.

Adjust 2_JOG-_CNEN-Error.py