Work around DLY STOP problem
When the motor is stopped (to be more exact: the motorrecord.STOP
is written with 1) while the motor record is in the DLY phase,
the motor record may lock up, when the motor position redback reports
the same postion all the time, and the motor record is now longer processed.
(This is easy to reproduce on a simulator. It may be difficult or impossible
to reproduce the on a real motor)

Work around this, by toggling the postion with +1.0 whenever the motor is
not moving.

Note: This works around another problem as well:
The motor record may stay unprocessed (UDF) when the IOC is powered
up and the motor position is exactly 0.

For more information see:
http://www.aps.anl.gov/epics/tech-talk/2015/msg01800.php