Send MRES to MCU when connected
Use MRES to calculate settings for the MCU in index group 5000, offset 0x23 and 024.
Scale MRES to be > 10.
Use %g instead of %f, since %f only prints 6 digits.
Changed the disconnection/connection handling:
Try to read out the axis index, and if that succeeds, synchronize MRES if needed.
Note: When the amplifier is on, MRES can not be changed in the MCU.
To avoid this, read the MCU settings first, and skip the update if they match.
If they don't match, indicate error 0x2001C.
The user can switch if the amplifier by setting CNEN to 0, and then MRES will be updated.

After the update, the user may enable the amplifier again.

Note: Changing MRES should happen rarely in practice, so the amplifier off restriction is OK.