Add ErrMsg
Make possible to send an Error Message as a string from the MCU to EPICS.

Add a PV like "$(R)ErrMsg" in the EPICS database.

Implement the handling in eemcuAxis.cpp:
Whenever nErrorID is != 0, retrieve sErrorMessage from the MCU and put it
into the database using setStringParam().

Whenever the MCU is disconnected from the IOC, display the dummy message
"Communication error"

Add code in the MCU for sErrorMessage:
For now, do a simple handling: Convert nErrorID into ASCII.