Don't send MRES from EPICS to the MCU
MRES in EPICS has 2 purposes:
a) It tells the motor record how to convert engineering units into steps
b) It defines the minimal possible movement

a) is done in the MCU and how this is done can be controlled by ADR commands.
What exactly needs to be put into the ADR commands is dependent on the motor, the encoder
and if an encoder is used at all.

Remove the synchronization of MRES and the MCU.