Add setADRValueOnAxisVerify() and use in in readConfigFile()
When a config file is read, it may be impossible to set ADR
values like scaling when the amplifier power is on.

Be more definsive:
Check the value on the axis first, and if it is already as
it should be, skip the update.

Introduce  setADRValueOnAxisVerify() both for int and double,
and use them in readConifFile().