simulator: Correct status when hitting LS
Problem:
When jogging and hitting a limit switch, the active limit switch
was reported in the next poll.

Solution:
Whenever stAxisStatus is asked, first the "simulated" motion
needs to be done.
The simultated motion calculates the new position,
CLIPS it if neccessary.
After that the status of the limit switches may be retrieved with
e.g. getPosLimitSwitch(()